# docs/swagger-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/arc-api/v1/projects/{projectId}/module-instances/get": {
      "post": {
        "operationId": "ModuleInstanceController_getModuleInstances",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the project",
            "schema": {
              "example": "12345",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "List of module instance system ids",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemIdsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModuleInstanceDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Some module instances are not found"
          },
          "422": {
            "description": "Failed to get module instances"
          }
        },
        "summary": "Get module instances for provided systemIds",
        "tags": [
          "module-instances"
        ]
      }
    },
    "/arc-api/v1/projects/{projectId}/module-instances": {
      "post": {
        "operationId": "ModuleInstanceController_addModuleInstance",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the project",
            "schema": {
              "example": "12345",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Request body parameter type: BaseModuleInstanceRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaseModuleInstanceRequest"
              },
              "examples": {
                "default": {
                  "summary": "Example BaseModuleInstanceRequest",
                  "value": {
                    "moduleId": 117444629,
                    "procId": 2,
                    "parentId": 251658241,
                    "subgraphId": 2952790017,
                    "containerId": 3758096385
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New created module information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleInstanceDto"
                },
                "examples": {
                  "default": {
                    "value": {
                      "_systemId": "1",
                      "_id": 1,
                      "_name": "Example Module",
                      "_relatedEndPointLinks": [
                        {
                          "_hypertextRef": "/components/1/properties",
                          "_method": "GET",
                          "_description": "Get properties for this module instance."
                        }
                      ],
                      "_editType": "None",
                      "_parentId": 202,
                      "_dataPorts": [],
                      "_controlPorts": [],
                      "alias": "ExampleAlias",
                      "moduleId": 123,
                      "subgraphId": 456,
                      "containerId": 789,
                      "maxInputPortsSupported": 5,
                      "maxOutputPortsSupported": 3,
                      "maxControlPortsSupported": 2,
                      "heapId": 101
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Failed to add a new module"
          }
        },
        "summary": "Create a new module instance for a given module id",
        "tags": [
          "module-instances"
        ]
      }
    },
    "/arc-api/v1/projects/{projectId}/module-instances/{moduleInstanceSystemId}/properties": {
      "get": {
        "operationId": "ModuleInstanceController_getModuleInstanceProperties",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the project",
            "schema": {
              "example": "12345",
              "type": "string"
            }
          },
          {
            "name": "moduleInstanceSystemId",
            "required": true,
            "in": "path",
            "description": "System id of a module instance",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleInstancePropertiesDto"
                }
              }
            }
          },
          "404": {
            "description": "Module instance is not found"
          },
          "422": {
            "description": "Failed to get module instance properties"
          }
        },
        "summary": "Get all property data for a module instance",
        "tags": [
          "module-instances"
        ]
      }
    },
...
    "/arc-api/v1/projects/{projectId}/data-links/get": {
      "post": {
        "operationId": "DataLinkController_getDataLinks",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the project",
            "schema": {
              "example": "12345",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "List of data-link system ids",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemIdsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataLinkDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Some data-link(s) are not found"
          },
          "422": {
            "description": "Failed to get data-link(s)"
          }
        },
        "summary": "Get data-links for provided systemIds",
        "tags": [
          "data-links"
        ]
      }
    },
    "/arc-api/v1/projects/{projectId}/data-links": {
      "post": {
        "operationId": "DataLinkController_addDataConnection",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the project",
            "schema": {
              "example": "12345",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Request body parameter type: NewDataLinkRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewDataLinkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataLinkDto"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Failed to add a data-link"
          }
        },
        "summary": "Add a new data link",
        "tags": [
          "data-links"
        ]
      }
    },
...
  },
  "info": {
    "title": "AudioReach Creator API",
    "description": "Audio Reach Creator API is an open source and cross platform backend framework for updating Audio Reach database files.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "EndPointLink": {
        "type": "object",
        "properties": {
          "hypertextRef": {
            "type": "string",
            "description": "Hypertext reference URL following 'project{projectId}'"
          },
          "method": {
            "type": "string",
            "description": "HTTP method",
            "example": "GET"
          },
          "description": {
            "type": "string",
            "description": "Description of the endpoint"
          }
        },
        "required": [
          "hypertextRef",
          "method",
          "description"
        ]
      },
...
      "DataPortDto": {
        "type": "object",
        "properties": {
          "systemId": {
            "type": "string",
            "description": "System ID"
          },
          "changeId": {
            "type": "string",
            "description": "Change ID"
          },
          "editType": {
            "type": "string",
            "description": "Edit type",
            "enum": [
              "Added",
              "Updated",
              "Deleted",
              "None"
            ]
          },
          "id": {
            "type": "object",
            "description": "Component ID"
          },
          "name": {
            "type": "string",
            "description": "Component name"
          },
          "componentType": {
            "type": "string",
            "description": "Component type",
            "enum": [
              "Module",
              "Subsystem",
              "Container",
              "DataLink",
              "ControlLink",
              "Subgraph"
            ]
          },
          "relatedEndPointLinks": {
            "description": "Related endpoint links",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EndPointLink"
            }
          },
          "portIoType": {
            "type": "string",
            "description": "Port IO type",
            "enum": [
              "Input",
              "Output"
            ]
          },
          "portType": {
            "type": "string",
            "description": "Port type",
            "enum": [
              "Static",
              "Dynamic"
            ]
          }
        },
        "required": [
          "systemId",
          "changeId",
          "editType",
          "id",
          "name",
          "componentType",
          "relatedEndPointLinks",
          "portIoType",
          "portType"
        ]
      },
      "ControlPortIntentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Intent ID"
          },
          "name": {
            "type": "string",
            "description": "Intent name"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "ControlPortDto": {
        "type": "object",
        "properties": {
          "systemId": {
            "type": "string",
            "description": "System ID"
          },
          "changeId": {
            "type": "string",
            "description": "Change ID"
          },
          "editType": {
            "type": "string",
            "description": "Edit type",
            "enum": [
              "Added",
              "Updated",
              "Deleted",
              "None"
            ]
          },
          "id": {
            "type": "object",
            "description": "Component ID"
          },
          "name": {
            "type": "string",
            "description": "Component name"
          },
          "componentType": {
            "type": "string",
            "description": "Component type",
            "enum": [
              "Module",
              "Subsystem",
              "Container",
              "DataLink",
              "ControlLink",
              "Subgraph"
            ]
          },
          "relatedEndPointLinks": {
            "description": "Related endpoint links",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EndPointLink"
            }
          },
          "portType": {
            "type": "string",
            "description": "Port type",
            "enum": [
              "Static",
              "Dynamic"
            ]
          },
          "controlPortName": {
            "type": "string",
            "description": "Control port name"
          },
          "intents": {
            "description": "Control port intents",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ControlPortIntentDto"
            }
          }
        },
        "required": [
          "systemId",
          "changeId",
          "editType",
          "id",
          "name",
          "componentType",
          "relatedEndPointLinks",
          "portType",
          "controlPortName",
          "intents"
        ]
      },
      "ModuleInstanceDto": {
        "type": "object",
        "properties": {
          "systemId": {
            "type": "string",
            "description": "System ID"
          },
          "changeId": {
            "type": "string",
            "description": "Change ID"
          },
          "editType": {
            "type": "string",
            "description": "Edit type",
            "enum": [
              "Added",
              "Updated",
              "Deleted",
              "None"
            ]
          },
          "id": {
            "type": "object",
            "description": "Component ID"
          },
          "name": {
            "type": "string",
            "description": "Component name"
          },
          "componentType": {
            "type": "string",
            "description": "Component type",
            "enum": [
              "Module",
              "Subsystem",
              "Container",
              "DataLink",
              "ControlLink",
              "Subgraph"
            ]
          },
          "relatedEndPointLinks": {
            "description": "Related endpoint links",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EndPointLink"
            }
          },
          "parentId": {
            "type": "object",
            "description": "Parent component ID"
          },
          "dataPorts": {
            "description": "Data ports",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataPortDto"
            }
          },
          "controlPorts": {
            "description": "Control ports",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ControlPortDto"
            }
          },
          "alias": {
            "type": "string",
            "description": "Module alias"
          },
          "moduleId": {
            "type": "number",
            "description": "Module ID"
          },
          "subgraphId": {
            "type": "number",
            "description": "Subgraph ID"
          },
          "containerId": {
            "type": "number",
            "description": "Container ID"
          },
          "maxInputPortsSupported": {
            "type": "number",
            "description": "Maximum number of input ports supported"
          },
          "maxOutputPortsSupported": {
            "type": "number",
            "description": "Maximum number of output ports supported"
          },
          "maxControlPortsSupported": {
            "type": "number",
            "description": "Maximum number of control ports supported"
          },
          "heapId": {
            "type": "number",
            "description": "Heap ID"
          }
        },
        "required": [
          "systemId",
          "changeId",
          "editType",
          "id",
          "name",
          "componentType",
          "relatedEndPointLinks",
          "dataPorts",
          "controlPorts",
          "alias",
          "moduleId",
          "subgraphId",
          "containerId",
          "maxInputPortsSupported",
          "maxOutputPortsSupported",
          "maxControlPortsSupported",
          "heapId"
        ]
      },
      "SystemIdsRequestDto": {
        "type": "object",
        "properties": {
          "systemIds": {
            "description": "Array of system IDs",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "systemIds"
        ]
      },
...
      "DataLinkDto": {
        "type": "object",
        "properties": {
          "systemId": {
            "type": "string",
            "description": "System ID"
          },
          "changeId": {
            "type": "string",
            "description": "Change ID"
          },
          "editType": {
            "type": "string",
            "description": "Edit type",
            "enum": [
              "Added",
              "Updated",
              "Deleted",
              "None"
            ]
          },
          "id": {
            "type": "object",
            "description": "Component ID"
          },
          "name": {
            "type": "string",
            "description": "Component name"
          },
          "componentType": {
            "type": "string",
            "description": "Component type",
            "enum": [
              "Module",
              "Subsystem",
              "Container",
              "DataLink",
              "ControlLink",
              "Subgraph"
            ]
          },
          "relatedEndPointLinks": {
            "description": "Related endpoint links",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EndPointLink"
            }
          },
          "sourceId": {
            "type": "number",
            "description": "Source component ID",
            "example": 12345
          },
          "sourcePortId": {
            "type": "number",
            "description": "Source port ID",
            "example": 1
          },
          "destinationId": {
            "type": "number",
            "description": "Destination component ID",
            "example": 67890
          },
          "destinationPortId": {
            "type": "number",
            "description": "Destination port ID",
            "example": 2
          },
          "parentId": {
            "type": "number",
            "description": "Parent ID",
            "example": 54321
          },
          "isDangling": {
            "type": "boolean",
            "description": "Is dangling",
            "example": false
          },
          "connectionType": {
            "type": "string",
            "description": "Connection type",
            "enum": [
              "MODULE_MODULE",
              "MODULE_SUBSYSTEM",
              "SUBSYSTEM_MODULE",
              "SUBSYSTEM_SUBSYSTEM"
            ],
            "example": "MODULE_MODULE"
          }
        },
        "required": [
          "systemId",
          "changeId",
          "editType",
          "id",
          "name",
          "componentType",
          "relatedEndPointLinks",
          "sourceId",
          "sourcePortId",
          "destinationId",
          "destinationPortId",
          "isDangling",
          "connectionType"
        ]
      },
...
      "BaseModuleInstanceRequest": {
        "type": "object",
        "properties": {
          "moduleId": {
            "type": "number",
            "description": "Module ID"
          },
          "procId": {
            "type": "number",
            "description": "Processor ID"
          },
          "parentId": {
            "type": "number",
            "description": "Parent ID, Optional"
          },
          "subgraphId": {
            "type": "number",
            "description": "Subgraph ID. Optional. If not provided, should create a new subgraph."
          },
          "containerId": {
            "type": "number",
            "description": "Container ID. Optional. If not provided, should create a container."
          }
        },
        "required": [
          "moduleId",
          "procId"
        ]
      },
...
      "BaseValueElement": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name property (e.g. \"enable\")"
          },
          "type": {
            "type": "string",
            "description": "Type property (e.g. \"uint32\", \"int16\", etc.)"
          },
          "value": {
            "type": "string",
            "description": "Value property (real value, e.g. 1, 2, 3, etc.)"
          },
          "valueLabel": {
            "type": "string",
            "description": "ValueLabel property (e.g. \"Enable\", etc.)"
          }
        },
        "required": [
          "name",
          "type",
          "value",
          "valueLabel"
        ]
      },
      "PropertyDto": {
        "type": "object",
        "properties": {
          "systemId": {
            "type": "string",
            "description": "System ID"
          },
          "propertyId": {
            "type": "number",
            "description": "Property ID"
          },
          "propertyName": {
            "type": "string",
            "description": "Property name"
          },
          "hasDefinition": {
            "type": "boolean",
            "description": "Has definition or not"
          },
          "propertyValues": {
            "description": "Property values",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BaseValueElement"
            }
          },
          "definitionLink": {
            "type": "object",
            "description": "Definition link"
          }
        },
        "required": [
          "systemId",
          "propertyId",
          "propertyName",
          "hasDefinition",
          "propertyValues"
        ]
      },
      "ModuleInstancePropertiesDto": {
        "type": "object",
        "properties": {
          "properties": {
            "description": "Array of module instance properties",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyDto"
            }
          }
        },
        "required": [
          "properties"
        ]
      },
...
      "NewDataLinkRequest": {
        "type": "object",
        "properties": {
          "startComponentId": {
            "type": "number",
            "description": "Start component ID",
            "example": 12345
          },
          "startPortId": {
            "type": "number",
            "description": "Start port ID",
            "example": 1
          },
          "endComponentId": {
            "type": "number",
            "description": "End component ID",
            "example": 67890
          },
          "endPortId": {
            "type": "number",
            "description": "End port ID",
            "example": 2
          },
          "parentId": {
            "type": "number",
            "description": "Parent ID",
            "example": 54321
          },
          "isDangling": {
            "type": "boolean",
            "description": "Is dangling",
            "example": false,
            "default": false
          }
        },
        "required": [
          "startComponentId",
          "startPortId",
          "endComponentId",
          "endPortId",
          "isDangling"
        ]
      },
...
    }
  }
}